Move `LANTERN_OF_TRACKING` to its own file
